Refactoring and updating imports.
