+++ zombpyg/agent.py
@@ -1,7 +1,7 @@
 import numpy, pygame
 import bisect
-from zombpyg.object import Bullet
-from zombpyg.sensor import Sensor
+from zombpyg.core.bullet import Bullet
+from zombpyg.core.sensor import Sensor
 from zombpyg.utils.geometry import _valid_angle, get_angle_and_distance_to_point, calculate_distance
 from zombpyg.utils.surroundings import Color
 from zombpyg.things import Player
