+++ zombpyg/core/things.py
@@ -2,11 +2,11 @@ import random
 import numpy
 import pygame
 
-from zombpyg.core import Thing, CircularThing, RectangularThing, FightingThing
-from zombpyg.sensor import Sensor
+# from zombpyg.core import Thing, CircularThing, RectangularThing, FightingThing # TODO
+from zombpyg.core.sensor import Sensor
 from zombpyg.utils.surroundings import Color, get_movement_estimates
 from zombpyg.utils.geometry import _valid_angle, get_angle_and_distance_to_point
-from zombpyg.weapons import ZombieClaws, Knife, Axe, Gun, Rifle, Shotgun
+from zombpyg.core.weapons import ZombieClaws, Knife, Axe, Gun, Rifle, Shotgun
 from zombpyg.actions import (
     MoveableThing,
     RotatableThing,
